+++ src/main/java/com/insightml/data/samples/decorators/LabelDecorator.java
@@ -68,7 +68,7 @@ public final class LabelDecorator<S extends Sample, E> extends AbstractDecorator
 	}
 
 	@Override
-	public double[][] features() {
+	public float[][] features() {
 		return ref.features();
 	}
