+++ src/main/java/com/insightml/models/DoubleModel.java
@@ -22,7 +22,7 @@ public interface DoubleModel extends ModelInfo {
 
 	double[] predictDouble(ISamples<? extends Sample, ?> instances);
 
-	double predict(double[] features, int[] featuresFilter);
+	double predict(float[] features, int[] featuresFilter);
 
 	int[] constractFeaturesFilter(final ISamples<? extends Sample, ?> instances);
